requirements.txt
openai
gradio==4.44.1
huggingface-hub==0.34.3
pydantic==2.10.6
python-dotenv
tiktoken